requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
docopt==0.6.2
idna==3.4
keyboard==0.13.5
pipreqs==0.4.11
prompt-toolkit==3.0.31
python-dotenv==0.21.0
requests==2.28.1
termcolor==2.0.1
urllib3==1.26.12
wcwidth==0.2.5
yarg==0.1.9
